+++ apps/client/src/shared/constants/text-limits.ts
